New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: test_wvrgcal   Put in Release Notes: No              Module(s): air_casawvr            Description: Got rid of compiler warnings about unused parameters 	   		and int/uint comparisons.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@35841 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/air_casawvr/cmdline/wvrgcal.cpp
@@ -230,7 +230,7 @@ LibAIR2::AntSet getAntPars(const std::string &s,
                           anames.end(), false,
 #if __cplusplus >= 201103L
                           [=](bool acc, const aname_t::value_type &p) {
-                              return acc || (p.first == n);
+			    return acc || ((int)p.first == n);
                           }
 #else
 						  hack02(n)
